Vector3 to vector4

+++ src/renderable_list.rs
@@ -3,10 +3,9 @@ use crate::{
     ray::Ray
 };
 use rand::Rng;
-use std::f64;
 
 pub struct Intersection {
-    pub t: f64,
+    pub t: f32,
     pub index: usize
 }
 
@@ -30,9 +29,9 @@ impl<R: Rng + ?Sized> RenderableList<R> {
     /// Finds the smallest value of `t` such that `r` intersects an element of the list and `t` lies in `[t_min, t_max]`, and the index `i`
     /// 
     /// of the list element that yields the minimal `t`. Returns `Some(Intersection { t, i })` if such a `t` is found, `None` otherwise.
-    pub fn intersect(&self, r: Ray, t_min: f64, t_max: f64) -> Option<Intersection> {
+    pub fn intersect(&self, r: Ray, t_min: f32, t_max: f32) -> Option<Intersection> {
         let mut intersection_index: usize = 0;
-        let mut intersection_t = f64::INFINITY;
+        let mut intersection_t = f32::INFINITY;
         for (i, e) in self.elements.iter().enumerate() {
             if let Some(t) = e.intersect(r, t_min, t_max) {
                 if t < intersection_t {
@@ -43,7 +42,7 @@ impl<R: Rng + ?Sized> RenderableList<R> {
                 continue;
             }
         }
-        // intersection_t should never be f64::INFINITY.
+        // intersection_t should never be f32::INFINITY.
         match intersection_t.is_finite() {
             true => Some(Intersection { t: intersection_t, index: intersection_index }),
             _ => None
